Split settings view descriptors
